requirements.txt
bokeh @ file:///Users/runner/miniforge3/conda-bld/bokeh_1625756994495/work
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1625835350093/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454858/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click @ file:///Users/runner/miniforge3/conda-bld/click_1621503850566/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494586/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1627691555247/work
distributed @ file:///Users/runner/miniforge3/conda-bld/distributed_1627703909534/work
dtw==1.4.0
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1626188337504/work
HeapDict==1.0.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1624926062656/work
locket==0.2.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1621455750847/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1620506033308/work
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1624947304325/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1626648422821/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
pandas==1.3.1
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1617910651905/work
patsy==0.5.1
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1625678050927/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1618863975118/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291460269/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
PyYAML==5.4.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1628432041960/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1624824941870/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1621217038088/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1612273721995/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1614001031686/work
tblib @ file:///home/conda/feedstock_root/build_artifacts/tblib_1616261298899/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1626092076920/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1600973991856/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1625488956802/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1627689047698/work
tsfresh @ file:///home/conda/feedstock_root/build_artifacts/tsfresh_1599727497869/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
zict==2.0.0
